pyproject.toml
[project]
name = "personal-docs-agent"
version = "0.1.1"
description = "Local-first personal docs RAG assistant with citations, ingestion, indexing, eval, and optional LoRA fine-tuning."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "numpy>=1.26.4",
  "fastembed>=0.3.6",
  "pypdf>=4.2.0",
  "PyMuPDF>=1.24.7",
  "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
web = [
  "fastapi>=0.110.0",
  "uvicorn>=0.30.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
]

[project.scripts]
docagent = "docagent.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docagent"]
include = [
  "src/docagent/web/templates/**",
  "src/docagent/web/static/**",
]
